requirements.txt
scikit-learn>=1.3.0
gensim>=4.3.0
bertopic>=0.16.0
sentence-transformers>=2.6.0
umap-learn>=0.5.3
hdbscan>=0.8.33
nltk>=3.8.1
pyLDAvis>=3.4.1
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
plotly>=5.18.0
wordcloud>=1.9.0
tqdm>=4.66.0
